convert swap_barval to double

+++ src/common.h
@@ -83,7 +83,7 @@ void print_swapmax(struct text_object *, char *, int);
 uint8_t mem_percentage(struct text_object *);
 double mem_barval(struct text_object *);
 uint8_t swap_percentage(struct text_object *);
-uint8_t swap_barval(struct text_object *);
+double swap_barval(struct text_object *);
 
 void print_kernel(struct text_object *, char *, int);
 void print_machine(struct text_object *, char *, int);
